Move some stuff from AbstractViewContainer to StaticUtils.
