+++ src/main/static/src/components/common/StaticUtils.js
@@ -31,6 +31,54 @@ class StaticUtils extends React.Component {
         a[Object.keys(b)[0]] = b[Object.keys(b)[0]];
         return a
     };
+
+    static findOrCreateYhteystiedotRyhmaFlat(henkiloUpdate, ryhmakuvaus, yhteystietotyyppi, label) {
+        let yhteystiedotRyhmaIndex = null;
+        let yhteystietoIndex = null;
+        let tyosahkopostiRyhma = henkiloUpdate.yhteystiedotRyhma
+            .filter((yhteystiedotRyhma, idx) => {
+                if(!yhteystiedotRyhmaIndex && yhteystiedotRyhma.ryhmaKuvaus === ryhmakuvaus) {
+                    yhteystiedotRyhmaIndex = idx;
+                    return true;
+                }
+                return false;
+            })[0];
+        let tyosahkoposti = tyosahkopostiRyhma
+            ? tyosahkopostiRyhma.yhteystieto.filter((yhteystieto, idx) => {
+                if(yhteystietoIndex === null && yhteystieto.yhteystietoTyyppi === yhteystietotyyppi) {
+                    yhteystietoIndex = idx;
+                    return true;
+                }
+                return false;
+            })[0]
+            : null;
+        if(yhteystiedotRyhmaIndex === null) {
+            yhteystiedotRyhmaIndex = henkiloUpdate.yhteystiedotRyhma.length;
+            tyosahkopostiRyhma = {
+                readOnly: false,
+                ryhmaAlkuperaTieto: "alkupera2", // Virkailija
+                ryhmaKuvaus: ryhmakuvaus,
+                yhteystieto: []
+            };
+            henkiloUpdate.yhteystiedotRyhma.push(tyosahkopostiRyhma);
+        }
+
+        if(yhteystietoIndex === null) {
+            yhteystietoIndex = henkiloUpdate.yhteystiedotRyhma[yhteystiedotRyhmaIndex].yhteystieto.length;
+            tyosahkoposti = {yhteystietoTyyppi: yhteystietotyyppi, yhteystietoArvo: ''};
+            henkiloUpdate.yhteystiedotRyhma[yhteystiedotRyhmaIndex].yhteystieto.push(tyosahkoposti);
+        }
+        return { label: label,
+            value: tyosahkoposti && tyosahkoposti.yhteystietoArvo,
+            inputValue: 'yhteystiedotRyhma.'+yhteystiedotRyhmaIndex+'.yhteystieto.'+yhteystietoIndex+'.yhteystietoArvo',
+        };
+    };
+
+    static createPopupErrorMessage(notificationKey, henkilo, L) {
+        const notification = henkilo.buttonNotifications[notificationKey];
+        return {errorTopic: notification && L[notification.notL10nMessage],
+            errorText: notification && L[notification.notL10nText]};
+    };
 }
 
 export default StaticUtils;
