+++ src/main/static/src/components/common/henkilo/buttons/MyonnaButton.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import AbstractViewContainer from "../../../../containers/henkilo/AbstractViewContainer";
 import ConfirmButton from "../../button/ConfirmButton";
+import StaticUtils from "../../StaticUtils";
 
 const MyonnaButton = ({L, myonnaAction, henkilo}) => <ConfirmButton action={myonnaAction}
                                             confirmLabel={L['HENKILO_KAYTTOOIKEUSANOMUS_MYONNA_CONFIRM']}
                                             normalLabel={L['HENKILO_KAYTTOOIKEUSANOMUS_MYONNA']}
                                             key="myonna"
-                                            errorMessage={AbstractViewContainer.createPopupErrorMessage('myonna', henkilo, L)} />;
+                                            errorMessage={StaticUtils.createPopupErrorMessage('myonna', henkilo, L)} />;
 
 MyonnaButton.propTypes = {
     L: React.PropTypes.object,
