+++ src/main/static/src/components/common/henkilo/buttons/PassivoiButton.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import ConfirmButton from "../../button/ConfirmButton";
 import Button from "../../button/Button";
-import AbstractViewContainer from "../../../../containers/henkilo/AbstractViewContainer";
+import StaticUtils from "../../StaticUtils";
 
 const PassivoiButton = ({henkilo, passivoiAction, L}) => henkilo.henkilo.passivoitu
     ? <Button key="passivoi" big disabled action={() => {}}>{this.L['PASSIVOI_PASSIVOITU']}</Button>
     : <ConfirmButton key="passivoi" big action={() => passivoiAction(henkilo.henkilo.oidHenkilo)}
                      normalLabel={L['PASSIVOI_LINKKI']} confirmLabel={L['PASSIVOI_LINKKI_CONFIRM']}
-                     errorMessage={AbstractViewContainer.createPopupErrorMessage('passivoi', henkilo, L)} />;
+                     errorMessage={StaticUtils.createPopupErrorMessage('passivoi', henkilo, L)} />;
 
 PassivoiButton.propTypes = {
     henkilo: React.PropTypes.shape({henkilo: React.PropTypes.shape({
