+++ src/main/static/src/components/common/henkilo/buttons/YksiloiHetutonButton.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import AbstractViewContainer from "../../../../containers/henkilo/AbstractViewContainer";
 import ConfirmButton from "../../button/ConfirmButton";
+import StaticUtils from "../../StaticUtils";
 
 const YksiloiHetutonButton = ({henkilo, L}) =>
     !henkilo.henkilo.yksiloityVTJ && !henkilo.henkilo.hetu
         ? <ConfirmButton key="yksilointi" big action={() => this.props.yksiloiHenkilo(this.props.henkilo.henkilo.oidHenkilo)}
                          normalLabel={L['YKSILOI_LINKKI']} confirmLabel={this.L['YKSILOI_LINKKI_CONFIRM']}
-                         errorMessage={AbstractViewContainer.createPopupErrorMessage('yksiloi', henkilo, L)} />
+                         errorMessage={StaticUtils.createPopupErrorMessage('yksiloi', henkilo, L)} />
         : null;
 
 YksiloiHetutonButton.propTypes = {
